+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 # Unreleased
 - Added MetaTable module
+- Added Pagination component
 
 ### Button
 - removed `value` property
-- `error` property renamed to `danger`
-- removed `bordered` property
 - removed `id` property
 - removed `iconLeft` and `iconRight` properties, use `icon` property instead
 - content of a button passed as a children instead of `text` property
@@ -12,7 +11,7 @@
 ### LinkButton
 - renamed to `Link`
 
-## ButtonIcon
+### ButtonIcon
 - removed, use Button instead
 
 # 0.2.0